requirements.txt
Flask==3.1.0
tensorflow==2.16.2
keras==3.9.2
numpy==1.26.4
Pillow==11.2.1
gdown==5.2.0
